requirements/requirements_hpc2n.txt
alabaster==0.7.12
appdirs==1.4.4
asn1crypto==1.4.0
atomicwrites==1.4.0
attrs==21.2.0
Babel==2.9.1
bcrypt==3.2.0
bitstring==3.1.7
blist==1.3.6
Bottleneck==1.3.2
CacheControl==0.12.6
cachy==0.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cleo==0.8.1
click==7.1.2
clikit==0.6.2
colorama==0.4.4
crashtest==0.3.1
cryptography==3.4.7
cycler==0.10.0
Cython==0.29.23
deap==1.3.1
decorator==5.0.7
distlib==0.3.1
docopt==0.6.2
docutils==0.17.1
ecdsa==0.16.1
expecttest==0.1.3
filelock==3.0.12
flit==3.2.0
flit-core==3.2.0
fsspec==2021.4.0
future==0.18.2
html5lib==1.1
idna==2.10
imagesize==1.2.0
importlib-metadata==4.0.1
iniconfig==1.1.1
intervaltree==3.1.0
intreehooks==1.0
ipaddress==1.0.23
jeepney==0.6.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
keyring==21.8.0
keyrings.alt==4.0.2
kiwisolver==1.3.1
liac-arff==2.5.0
lockfile==0.12.2
MarkupSafe==1.1.1
matplotlib==3.4.2
mock==4.0.3
more-itertools==8.7.0
mpmath==1.2.1
msgpack==1.0.2
netaddr==0.8.0
netifaces==0.10.9
nose==1.3.7
numexpr==2.7.3
numpy==1.20.3
packaging==20.9
pandas==1.2.4
paramiko==2.7.2
pastel==0.2.1
pathlib2==2.3.5
paycheck==1.0.2
pbr==5.6.0
pexpect==4.8.0
Pillow==8.2.0
pip==21.1.1
pkginfo==1.7.0
pluggy==0.13.1
poetry==1.1.6
poetry-core==1.0.3
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
py-expression-eval==0.3.13
pyasn1==0.4.8
pybind11==2.6.2
pycparser==2.20
pycrypto==2.6.1
Pygments==2.9.0
pylev==1.3.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.4
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-toolbelt==0.9.1
scandir==1.10.0
scipy==1.6.3
SecretStorage==3.3.1
semantic-version==2.8.5
setuptools==56.2.0
setuptools-rust==0.12.1
setuptools-scm==6.0.1
shellingham==1.4.0
simplegeneric==0.8.1
simplejson==3.17.2
six==1.16.0
snowballstemmer==2.1.0
sortedcontainers==2.3.0
Sphinx==4.0.0
sphinx-bootstrap-theme==0.7.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.2.4
tabulate==0.8.9
threadpoolctl==2.1.0
toml==0.10.2
tomlkit==0.7.0
torch==1.10.0
torchvision==0.11.1
tqdm==4.61.2
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.4
virtualenv==20.4.6
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.36.2
xlrd==2.0.1
zipp==3.4.1
openpyxl